Prints message using putchar

+++ 0x00-hello_world/101-quote.c
@@ -10,7 +10,7 @@ int main(void)
 	char message[] = "and that piece of art is useful \" - Dora Korpar, 2015-10-19";
 	int i;
 
-	for (i = 0; i < sizeof(message); i++)
+	for (i = 0; i < (int) sizeof(message); i++)
 	{
 		putchar(message[i]);
 	}
